### 2020-08-30 #2613 @whedon
[PRE REVIEW]: Moran Pycess: a Python package to simulate Moran processes driven by game theory
**Submitting author:** @AngryMaciek (<a href="http://orcid.org/0000-0003-1361-7301">Maciej Bak</a>)
**Repository:** <a href="https://github.com/AngryMaciek/angry-moran-simulator" target ="_blank">https://github.com/AngryMaciek/angry-moran-simulator</a>
**Version:** v1.0.0
