### 2020-09-04 #2613 @majensen
@spond - it has been a long time! Would you be willing to perform a software review of [this work](#2613) for the [Journal of Open Source Software](https://joss.theoj.org)? All reviews are performed openly and online in GitHub; an overview of the expectations are [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

It would be great to have your popgen expertise on this work. If you have any suggestions for others, that would be great too. Please reply in this issue thread. Thanks for your consideration -- @majensen
