### 2020-09-04 #2613 @majensen
@RHagenson - reaching out to see if you would have time and interest in reviewing [this work](#2613) for [JOSS](https://joss.theoj.org). Thanks for your consideration!
