### 2020-09-06 #2613 @majensen
@xin-huang and @tkchafin Thanks very much for stepping up -- I agree with @rozmanowna that you both would bring excellent experience to this review. Please refresh your memories regarding the [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I will assign you as reviewers and then begin the review. The review proper will take place in a separate GitHub issue.
Much apppreciated!
