### 2020-09-07 AngryMaciek/angry-moran-simulator#5 @xin-huang
[JOSS REVIEW] Suggestions for Documentation
This is a part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/2643)

1. I found a Jupyter notebook https://github.com/AngryMaciek/angry-moran-simulator/blob/master/tests/usecase.ipynb could reproduce the examples. I suggest the authors provide a link to this notebook in their documentation explicitly, so that users can quickly replicate the examples.
2. After installing the package successfully in my computer with Ubuntu 18.04, I failed to replicate the examples in https://github.com/AngryMaciek/angry-moran-simulator/blob/master/tests/usecase.ipynb at first. I have to manually install `jupyter`, `pyyaml` and `seaborn` under the virtual environment `moranpycess` before successfully running the notebook. Thus, I suggest the authors add `jupyter`, `pyyaml` and `seaborn` into https://github.com/AngryMaciek/angry-moran-simulator/blob/master/env/main.yml, so that users can easily reproduce the examples.
3. As the checklist suggests, a statement of need is required in the documentation. Besides, the authors may provide a link to https://github.com/AngryMaciek/angry-moran-simulator/blob/master/env/main.yml, so that users can find dependencies for this package easily.
4. Typos in https://github.com/AngryMaciek/angry-moran-simulator/blob/master/documentation.md
    - The word "sygnature" should be "signature"
