### 2020-09-07 AngryMaciek/angry-moran-simulator#6 @xin-huang
[JOSS REVIEW] Suggestions for Software Paper
This is a part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/2643)

1. As the checklist suggests, a Summary section is required in the paper.
2. Although the authors discussed some other simulators (DEAP and Axelrod) in the statement of need, I suggest the authors separate these discussion into a State of the field section, as the checklist suggests, or merge them into the Introduction section.
