### 2020-09-07 #2643 @xin-huang
The authors implemented a python package to perform simulation with different evolutionary strategies under Moran process. This is an interesting simulator in population biology, because there are very few user-friendly simulators for evolutionary game currently. After testing this package and reading the paper, I have several concerns on the documentation and software paper. I have opened some issues (https://github.com/AngryMaciek/angry-moran-simulator/issues/5 and https://github.com/AngryMaciek/angry-moran-simulator/issues/6) in the repository. Because I do not have computers with MacOS and TravisCI cannot make python builds on MacOS. I cannot verify the installation and performance on MacOS.
