### 2020-09-07 AngryMaciek/angry-moran-simulator#7 @tkchafin
[JOSS Review] Conda installation "Conflicts found"
This is a part of the JOSS review (openjournals/joss-reviews#2643)

Environment creation as suggested in the documentation yields a bunch of conflicts found by conda. I was able to amend this by relaxing the version constraints in the .yml file like so:

---

  name: moranpycess

  channels:
    - bioconda
    - conda-forge

  dependencies:
    - matplotlib=3.3
    - numpy=1.19
    - pandas
    - pip=20.1
    - python=3.7

...

Also, as suggested by @xin-huang, I would request that you add any additional dependencies needed for plotting, such as seaborn.
