### 2020-09-07 AngryMaciek/angry-moran-simulator#8 @tkchafin
[JOSS Review] Insufficient documentation for plotting functions.
This is a part of the JOSS review (openjournals/joss-reviews#2643)

Documentation in 'documentation.md' is not sufficient for the MoranProcess plotting methods. (PlotSize, PlotAvgBirthPayoff, PlotAvgDeathPayoff, PlotBirthFitness, PlotDeathFitness, PlotEntropy).Please amend by providing explicit examples.
