### 2020-09-07 AngryMaciek/angry-moran-simulator#9 @tkchafin
[JOSS Review] Writing quality suggestions for software paper
This is a part of the JOSS review (openjournals/joss-reviews#2643)

Some very minor suggestions to improve clarity of the software paper:

- ESS versus EES is not consistent. Please fix to the correct abbreviation (=ESS?)
- Intro L1: "Population is created by" would be more clearly written as "A population consists of individuals" or similar
- Intro L2-4: "Population characteristics, including: ..." would be more clearly written as something like "Population characteristics (e.g., reproduction, .... etc) change over time as a result of evolutionary mechanisms"
- Intro L4: "*A* complicated network of..."
- Intro L12: EES or ESS?
- Intro L13: should it be "evolutionarily" stable?
- Statement of need L1: Perhaps mention the name of the package in the first sentence?
- Statement of need L2: "Contrary to cellular automata"? Perhaps revise for clarity
- Statement of need L13: "An open source license"
- Example (A) L2: Define "player"
