### 2020-09-07 #2643 @tkchafin
The authors have provided a flexible and user-friendly software to generate simulations under several different scenarios from evolutionary game theory. I was able to verify functionality of the simulated processes and built-in plotting functions under MacOS, after minor changes to the documented installation instructions provided by the authors. After completing my review, I would recommend publication of this contribution in JOSS after completing several necessary revisions, which are aimed primarily at increasing clarity of the communications associated with this package (both software paper and documentation), and making installation/ plotting easier for the user. These are summarized below, and in greater detail as Issues on the software github page.

Software paper:
Major issues (per checklist) are currently: 1) The lack of a non-specialist summary statement; and 2) There are many inconsistencies in the writing that make me suggest the authors go back through more carefully to edit for clarity and language). For example, “EES” vs “ESS” is not consistent, and several other instances (i.e. “A population consists of individuals…” rather than “Population is created by individuals…”) where the quality of writing could be improved. These problems are listed in more detail over at issue #9(https://github.com/AngryMaciek/angry-moran-simulator/issues/9).

Documentation/ Functionality: There are 2 major issues here that I think should be corrected prior to publication: 1) The installation was not initially successful (on MacOS) without first manually relaxing the version constraints specified in the .yml file. I suspect that these constraints are overly specific, which for me created conflicts when creating the environment in conda; and 2) Documentation for the plotting functions is insufficient. A third issue (which I consider minor) is that - per the checklist - the documentation should also contain a statement of need. These are also added as issues #7 (https://github.com/AngryMaciek/angry-moran-simulator/issues/7) and #8 (https://github.com/AngryMaciek/angry-moran-simulator/issues/8)
