### 2020-09-08 AngryMaciek/angry-moran-simulator#5 @AngryMaciek
(4) should be fixed in: https://github.com/AngryMaciek/angry-moran-simulator/commit/c3b2f6aeb595d5f06575b81588f7da869835eedf
