### 2020-09-08 ef716cb
docs: add TOC to documentation.md, usecase link, move 1line to README.md
