### 2020-09-08 932967a
docs: remove duplicated info
