+++ documentation.md
@@ -1,6 +1,6 @@
 # Moran [Py]cess: Documentation & How-to
 
-- [Moran [Py]cess: Documentation & How-to](#moran-[py]cess:-documentation-&-how-to)
+- [Moran [Py]cess: Documentation & How-to](#moran-pycess-documentation--how-to)
   - [General Moran Process](#general-moran-process)
   - [Moran Model based on 2D neighbourhood](#moran-model-based-on-2d-neighbourhood)
   - [Moran Model based on 3D neighbourhood](#moran-model-based-on-3d-neighbourhood)
