### 2020-09-08 AngryMaciek/angry-moran-simulator#5 @AngryMaciek
(1) is fixed from 0fe50fb forward, also added a small Table Of Contents in the documentation.
