### 2020-09-08 1f1a09b
build: updated main env, removed pyyaml dependancy
