### 2020-09-08 AngryMaciek/angry-moran-simulator#5 @AngryMaciek
(2) is fixed from: 1f1a09b forward; `pyyaml` is not necessary so I removed the `import` statement. Please note that the environment will further change according to the suggestions in #7 .
