### 2020-09-08 cc8235e
ci: add macOS build
