### 2020-09-08 AngryMaciek/angry-moran-simulator#7 @AngryMaciek
Dear @tkchafin ,
I have added a build on Travis macOS machine to the CI in https://github.com/AngryMaciek/angry-moran-simulator/commit/cc8235e1f27071301b807f1025f1aa6d0f6c2c73, please take a look at the attached build logs:
https://travis-ci.org/github/AngryMaciek/angry-moran-simulator/builds/725193448
It did not yield any conda conflicts. I have also built the environment on my personal mac, with no conflicts either. Are you using an older conda version, perhaps?
