### 2020-09-08 2eaa894
docs: add dependencies info
