### 2020-09-08 AngryMaciek/angry-moran-simulator#5 @AngryMaciek
(3) is addressed in 71b6216: added two sections to the documentation file.

@xin-huang : Thank you for the comments! I believe I have addresses all of the suggestions above.
