### 2020-09-08 AngryMaciek/angry-moran-simulator#5 @xin-huang
Hi @AngryMaciek, thank you for addressing my concerns. All the changes look good.
