### 2020-09-08 3be60f5
docs: correct plotting functions description
