### 2020-09-08 AngryMaciek/angry-moran-simulator#8 @AngryMaciek
Dear @tkchafin,
Thanks a lot for the notice, indeed there were mistakes in the description of these functions.
In 3be60f5 I have introduced proper information about their execution.
Please take a look at the updated documentation.
