### 2020-09-09 ca0b710
paper: add State of the field section
