### 2020-09-09 ad477ff
paper: modify Introduction and Example results sections
