+++ paper.md
@@ -29,10 +29,7 @@ Moran Pycess is a bioinformatics Python package created to simulate population�
 
 # Introduction
 
-Population is created by individuals of the same species living simultaneously on a shared area and related to each other by a complex system of mutual dependencies.
-Population characteristics, including: reproduction, mortality, abundance or life strategy change over time under the influence of evolutionary mechanisms. Complicated network of interactions with several degrees of freedom is very challenging for evolutionary biologists to decode. A simple mathematical model proposed by Patrick Moran is often used to describe probabilistic dynamics of a finite population of constant size [@Moran:1958]. Within this framework, each individual might be assigned with a fitness score calculated solely based on the scored interactions with all the other members of the group. Fitness drives the probability of an individual to reproduce while mutual interaction scores are calculated according to a common ‘payoff matrix’, treating the whole system as a game. In many populations the emergence of evolutionary stable strategy (EES) is observed. ESS complements Nash's equilibrium with an additional stability condition. A strategy is considered as evolutionary stable if it is resistant to an invasion of a small group with a different phenotype and cannot be overtaken [@Smith:1982]. Evolutionary game theory set grounds 
-for a solid framework for quantitative population biology, allowing researchers 
-to simulate dynamics and estimate trajectories of biological systems.
+A population consists of individuals of the same species living simultaneously on a shared area and related to each other by a complex system of mutual dependencies. Population characteristics (e.g reproduction, mortality, abundance or life strategy) change over time as a result of evolutionary mechanisms. A complicated network of interactions with several degrees of freedom is very challenging for evolutionary biologists to decode. A simple mathematical model proposed by Patrick Moran is often used to describe probabilistic dynamics of a finite population of constant size [@Moran:1958]. Within this framework, each individual might be assigned with a fitness score calculated solely based on the scored interactions with all the other members of the group. Fitness drives the probability of an individual to reproduce while mutual interaction scores are calculated according to a common ‘payoff matrix’, treating the whole system as a game. In many populations the emergence of evolutionary stable strategy (ESS) is observed. ESS complements Nash's equilibrium with an additional stability condition. A strategy is considered evolutionarily stable if it is resistant to an invasion of a small group with a different phenotype and cannot be overtaken [@Smith:1982]. Evolutionary game theory set grounds for a solid framework for quantitative population biology, allowing researchers to simulate dynamics and estimate trajectories of biological systems.
 
 # Statement of need 
 
@@ -48,13 +45,10 @@ We have designed four distinct systems based on well-known interaction examples 
 
 ![Simulations of population evolution based on: (A) Stag Hunt (B) Chicken game, (C) Prisoners' Dilemma, (D) Rock Paper Scissors.](images/figure.png)
 
-(A) The stag hunt model describes a conflict between safety and social cooperation: hare -  small but certain profit, stag - great benefit but adventurous. Failure in cooperation leads to a player’s loss. There is no dominant strategy in this game as it is most beneficial for individuals to agree on the same strategy. Both stag hunt and hare hunt are considered as an ESS. A perfect example of this type of interaction is the “carousel feeding” as a cooperative hunting method used by Norwegian orcas which force the school of fish to form a tight ball, then stun fish with their tails [@Fort:2007].
-
[…]
