+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-Moran Pycess is a bioinformatics Python package created to simulate population’s growth based on evolutionary game theory and the Moran Process model. Growth dynamics is controlled solely by the interaction of individuals with each other. Our software is primarily targeted to evolutionary and computational biologists, however students who learn basic concepts behind population’s evolution may find it equally practical.
+Moran Pycess is a bioinformatics Python package created to simulate population’s growth based on evolutionary game theory and the Moran Process model. Growth dynamics is controlled solely by the interactions of individuals with each other. Our software is primarily targeted to evolutionary and computational biologists, however students who learn basic concepts behind population’s evolution may find it equally practical.
 
 # Introduction
 
