### 2020-09-09 AngryMaciek/angry-moran-simulator#6 @rozmanowna
Thank you for the comments @xin-huang!

I added Summary and State of the field sections as you suggested.

Here is the link to new pdf paper: https://github.com/openjournals/joss-reviews/issues/2643#issuecomment-689807532
