### 2020-09-09 AngryMaciek/angry-moran-simulator#9 @rozmanowna
@tkchafin thank you for the comments!

I hope I have addressed all of your suggestions.

I wonder about the concept of "player": if we consider abstract models of conflict situations like Prisoners' Dilemma - in this context, the concept of a player seems to be as basic as the concept of the "game" itself and it is hard to find a clear definition. It is simply a participant in the game and in our eyes it does not require further specification. To make it clearer in the text as such I have added "game" in Example (A) L1 in the text. I hope you find this sufficient?

Here is the link to new pdf paper: https://github.com/openjournals/joss-reviews/issues/2643#issuecomment-689807532
