### 2020-09-10 AngryMaciek/angry-moran-simulator#6 @xin-huang
Hi @rozmanowna, thanks for revising the paper.
But I still feel confused about this sentence

`Notably, another powerful Python package - Axelrod - allows to model evolution of cooperation in various strategies, but unlike Moran Pycess is limited to Prisoners’Dilemma only (The Axelrod project developers, 2016)`

It looks like the exact meaning should be
`Notably, another powerful Python package - Axelrod - is limited to Prisoners' Dilemma, while Moran Pycess allows to model evolution of cooperation in various strategies`

Another concern is "population's evolution" in the Summary section.  I feel it is redundant to say population's evolution. I think simply using evolution is OK.
