### 2020-09-10 26f121c
paper: midify Summary and State of teh field sections
