### 2020-09-10 c9250c3
paper: modify Summary Section
