### 2020-09-10 AngryMaciek/angry-moran-simulator#6 @rozmanowna
Thank you @xin-huang, I have addressed your additional suggestions.

Here is the link to new pdf paper: https://github.com/openjournals/joss-reviews/issues/2643#issuecomment-690122031
