### 2020-09-10 AngryMaciek/angry-moran-simulator#6 @xin-huang
Thanks @rozmanowna, I have no further concerns.
