### 2020-09-10 #2643 @majensen
Thanks @xin-huang  - so @AngryMaciek I will look for https://github.com/AngryMaciek/angry-moran-simulator/issues/7 to be addressed and @tkchafin to then tick the final box on that review. I'll then proceed with editorial checks. Thanks all!
