### 2020-09-13 AngryMaciek/angry-moran-simulator#7 @AngryMaciek
Dear @tkchafin
Are there any other modifications you would like me to implement?
