### 2020-09-23 AngryMaciek/angry-moran-simulator#7 @majensen
@tkchafin can you comment here or back at https://github.com/openjournals/joss-reviews/issues/2643 ? Thanks very much -- @majensen
