### 2020-09-23 AngryMaciek/angry-moran-simulator#7 @tkchafin
Hi sorry about the delay- in the middle of a move. @AngryMaciek either explicitly list dependencies in README/ docs or relax the version constraints in the yaml file. Ideally both, or give some justification why such specific version constraints are needed.
