### 2020-09-23 df07b1a
docs: update README ver. info
