### 2020-09-23 AngryMaciek/angry-moran-simulator#7 @AngryMaciek
OK, I have elaborted a little the `installation` section in the README by providing links to the *YAML* files which state the versions of the dependancies.

To care about *reproducible research* you generally want to be as specific as possible. You cannot guarantee that your code will work with many possible versions - you can guarantee just the one you are testing it with. So your code may be OK with `dependency1=0.1.0` but break on `dependency1=0.1.4`, who knows? Theoretically... This should not happen with patch versions updates but we cannot count that all the other developers stick to proper semantic versioning.
