### 2020-09-27 #2643 @majensen
@openjournals/dev - please see error in compilation at https://github.com/openjournals/joss-reviews/issues/2643#issuecomment-699695148 - I was able to use https://whedon.theoj.org successfully.
