### 2020-09-27 #2643 @majensen
[Proof generated by whedon.theoj.org](https://github.com/openjournals/joss-reviews/files/5289179/sp4qyhqedxejp6q3hoho.pdf)
