### 2020-10-01 #2643 @AngryMaciek
Dear Editors,
Should we now make a `zenodo` submission ourselves, or... (?)
