### 2020-10-02 #2643 @majensen
You can do that if you wish @AngryMaciek. I will have a pull request for you with a number of suggested changes. Apologies for the delay, thanks for your patience.
