### 2020-10-09 446036f
(ed) Maynard Smith ref.
Maynard Smith is JMS' surname
