+++ paper.bib
@@ -11,7 +11,7 @@ doi = {10.1017/S0305004100033193}
 }
 
 @book{Smith:1982, 
-author = {{Smith}, John Maynard},
+author = {{Maynard Smith}, John},
 title = {Evolution and the Theory of Games},
 publisher = {Cambridge University Press}, 
 year = {1982},   
