### 2020-10-09 5245340
(ed) clarifying and smoothing edits
The the yeast example is not clearly described and I think is missing
some of the key biological aspects of that system. It should be
rewritten, or another (simpler) example provided.
