+++ paper.md
@@ -29,26 +29,26 @@ Moran Pycess is a bioinformatics Python package created to simulate population�
 
 # Introduction
 
-A population consists of individuals of the same species living simultaneously on a shared area and related to each other by a complex system of mutual dependencies. Population characteristics (e.g reproduction, mortality, abundance or life strategy) change over time as a result of evolutionary mechanisms. A complicated network of interactions with several degrees of freedom is very challenging for evolutionary biologists to decode. A simple mathematical model proposed by Patrick Moran is often used to describe probabilistic dynamics of a finite population of constant size [@Moran:1958]. Within this framework, each individual might be assigned with a fitness score calculated solely based on the scored interactions with all the other members of the group. Fitness drives the probability of an individual to reproduce while mutual interaction scores are calculated according to a common ‘payoff matrix’, treating the whole system as a game. In many populations the emergence of evolutionary stable strategy (ESS) is observed. ESS complements Nash's equilibrium with an additional stability condition. A strategy is considered evolutionarily stable if it is resistant to an invasion of a small group with a different phenotype and cannot be overtaken [@Smith:1982]. Evolutionary game theory set grounds for a solid framework for quantitative population biology, allowing researchers to simulate dynamics and estimate trajectories of biological systems.
+A population consists of individuals of the same species living simultaneously on a shared area and related to each other by a complex system of mutual dependencies. Population characteristics (e.g reproduction, mortality, abundance or life strategy) change over time as a result of evolutionary mechanisms. A complicated network of interactions with several degrees of freedom is very challenging for evolutionary biologists to decode. A simple mathematical model proposed by Patrick Moran is often used to describe probabilistic dynamics of a finite population of constant size [@Moran:1958]. Within this framework, each individual might be assigned with a fitness score calculated solely based on the scored interactions with all the other members of the group. Fitness drives the probability of an individual to reproduce while mutual interaction scores are calculated according to a common ‘payoff matrix’, treating the whole system as a game. In many populations the emergence of evolutionary stable strategy (ESS) is observed. ESS complements Nash's equilibrium with an additional stability condition. A strategy is considered evolutionarily stable if it is resistant to an invasion of a small group with a different phenotype and cannot be supplanted [@Smith:1982]. Evolutionary game theory set grounds for a solid framework for quantitative population biology, allowing researchers to simulate dynamics and estimate trajectories of biological systems.
 
 # Statement of need 
 
-We have developed Moran Pycess - a Python package with a general game-theoretical framework for scientific simulations according to the Moran model. Such an approach allows the stochastic nature of evolution to be preserved. This is not the case for cellular automata, which are completely deterministic systems and where evolution of each cell follows some strictly defined local rules. We chose Python because of its availability and popularity in the fields of bioinformatics and data analysis. It is worth emphasizing that Moran Pycess is capable of carrying out simulations over 2D and 3D grid where individuals consider their direct neighbours. Three-dimensional space is particularly important for modeling dynamics of population growth in cell biology [@Macnamara:2020]. Another strength of Moran Pycess lays in its simplicity which turns it into a useful research aid for evolutionary and computational biologists. An open source license as well as its accessibility endorse Moran Pycess as a practical tool for biology, economics and math students to learn about population evolution based on game theory or for computer science students aiming to properly encapsulate their research software. A remarkable advantage of our module is that any possible model of an antagonistic game may be considered. In terms of quality assurance: our repository incorporates Travis CI mechanism alongside Coveralls code coverage measurement (currently: 100%).
[…]
