### 2020-10-09 #2643 @majensen
@AngryMaciek please have a look at the suggestions in the PR (and assoc. comments).

I would also like to see an example run in the paper that exemplifies stochastic effects. One example might be to show that if an invading population is small enough, it can disappear frequently (i.e., among many replicate simulations) even if it uses an ESS. Probably (absolute guess) the fitness benefit has to exceed 1/N more or less to get a high probability of establishment.
(This is also a test of the software - if a rare invader never disappears over multiple replicates, there is probably a code issue.)
thanks
