### 2020-10-09 16c5539
paper: add hawk-dove info
