### 2020-10-09 8358b10
paper: add hawk-dove caption
