### 2020-10-15 bd8070e
modify Example results - example B
