+++ paper.md
@@ -46,7 +46,7 @@ We have designed four distinct systems based on well-known interaction examples 
 ![Simulations of population evolution based on: (A) Stag Hunt (B) Chicken game, (C) Prisoners' Dilemma, (D) Rock Paper Scissors.](images/figure.png)
 
 (A) The stag hunt game model describes a conflict between safety and social cooperation: hare - small but certain profit, stag - great benefit but adventurous. Failure in cooperation leads to a player’s loss. There is no dominant strategy in this game as it is most beneficial for individuals to agree on the same strategy. Both stag hunt and hare hunt are considered as an ESS. A perfect example of this type of interaction is the “carousel feeding” as a cooperative hunting method used by Norwegian orcas which force the school of fish to form a tight ball, then stun fish with their tails [@Fort:2007].
-(B) In the game of Chicken a confrontational strategy brings the greatest profit. However, if chosen by both individuals it yields the worst outcome. A peace strategy protecting against the greatest loss brings no reward. No player has a dominant strategy. There is no ESS in pure strategies in this game. Consider an example of two strains of the yeasts S. cerevisiae: wildtype - using invertase to produce monosaccharides and a mutant lacking the invertase gene - consuming glucose without bearing the metabolic cost. Under certain conditions wildtype cells competition against always-defecting cells results in wildtype cells cooperation in invertase expressing. Invertase expression is repressed while competing against an always-cooperating strain causing the wildtype cell to cheat [@Gore:2009].
+(B) In the game of Chicken a confrontational strategy brings a greater profit. However, if chosen by both individuals it yields the worst outcome. A peaceful strategy, while protecting against the greatest possible loss brings no reward when challenged by an aggressor. No player has a dominant strategy and there are no ESS in pure strategies in this game. Consider an example of two strains of the yeasts S. cerevisiae: wildtype and a mutant lacking the invertase gene. In a monosaccharide-poor environment yeasts may utilise invertase to produce an energy source from more complex molecules. Most monosaccharides are excreted to the extracellular environment from where they can also be taken up by other cells. Mutant strains (“cheaters”) follow a confrontational strategy: they consume glucose created by wildtype cells without bearing the metabolic cost of invertase production and secretion. Therefore, in a population, a confrontational strategy has a good advantage over the peaceful one, given the number of wildtype yeasts exceeds the number of mutants vastly. In case this condition is not met cheaters would fail to live off the others, leading to starvation - the worst possible scenario for them. Moreover, the analogy holds while inspecting the behaviour of wildtype yeasts. Trying to avoid that worst scenario, in the presence of mutants, wildtype cells cooperate in invertase expression over a wide range of conditions. However, when competing against other wildtype strains, invertase expression is repressed, causing them to exhibit the confrontational strategy - cheat. [@Gore:2009].
 (C) Prisoners’ dilemma, often analyzed in game theory due to its multiple applications, presents a case where two completely rational individuals might not cooperate, even if it appears that it is in their best interests to do so. The most rational strategy yields the worst collective outcome. Every participant has a dominant strategy. The only ESS is to always cooperate. In biological context the prisoner’s dilemma successfully captures the essential features of cancer growth and allows to test hypotheses and formulate claims in a quantitative manne [@West:2016]. In the case of malignant tumors, we observe the possibility of infiltration of adjacent tissues by cancer cells. Their characteristic uncontrolled proliferation leads to the growth of another tumor in this place and the destruction of healthy tissue.
[…]
