### 2020-10-15 #2643 @AngryMaciek
We have just included the most recent changes in the branch `paper` of the [repository](https://github.com/AngryMaciek/angry-moran-simulator).
