### 2020-10-21 #2643 @majensen
@AngryMaciek I think things look quite good. Can you now create your archive, making sure that the archive title and authors are the same as those on the paper? Once that's done, please provide the archive DOI in this thread. Then I will attempt the spell summoning the editors-in-chief. thanks!
