### 2020-10-21 #2643 @AngryMaciek
@majensen

We have merged branch `paper` to `master` and removed the former.
We have submitted the code to zenodo: https://doi.org/10.5281/zenodo.4114143

Good luck with the spellcasting!
(∩｀-´)⊃━☆ﾟ.*･｡ﾟ
