### 2020-10-25 #2643 @kyleniemeyer
Hi @AngryMaciek, I'm the EIC on duty this week doing some final checks. I only noticed one minor issue: a typo in "e.g" in the second paragraph. Could you fix with "e.g.,"? Thanks!
