### 2020-10-25 #2643 @AngryMaciek
Hi @kyleniemeyer :)
Done in https://github.com/AngryMaciek/angry-moran-simulator/commit/431773d1872301f62dbdde76283988cf0fa8797e
