### 2020-10-26 #2643 @majensen
@AngryMaciek congrats- It's a nice application and an interesting paper. Thanks for choosing JOSS!
